passing down device type and removing tedge util
