+++ crates/core/tedge/src/cli/connect/command.rs
@@ -118,12 +118,15 @@ impl Command for ConnectCommand {
             );
         self.config_repository.store(&config)?;
 
+        let device_type = config.query(DeviceTypeSetting)?;
+
         new_bridge(
             &bridge_config,
             &updated_mosquitto_config,
             self.service_manager.as_ref(),
             self.user_manager.clone(),
             &self.config_location,
+            &device_type,
         )?;
 
         match self.check_connection(&config) {
@@ -373,6 +376,7 @@ fn new_bridge(
     service_manager: &dyn SystemServiceManager,
     user_manager: UserManager,
     config_location: &TEdgeConfigLocation,
+    device_type: &str,
 ) -> Result<(), ConnectError> {
     println!("Checking if {} is available.\n", service_manager.name());
     let service_manager_result = service_manager.check_operational();
@@ -394,8 +398,6 @@ fn new_bridge(
 
     if bridge_config.cloud_name.eq("c8y") {
         println!("Creating the device in Cumulocity cloud.\n");
-        let tedge_config = tedge_config()?;
-        let device_type = tedge_config.query(DeviceTypeSetting)?;
         let () = c8y_direct_connection::create_device_with_direct_connection(
             user_manager,
             bridge_config,
